refactor: 마이펫 페이지 코드 분리 및 스크롤 디자인 적용
