+++ src/components/my-pet/Heading.jsx
@@ -1,0 +1,11 @@
+import { styled } from "styled-components";
+
+export const Heading = styled.p`
+  margin: 8px 0;
+  font-size: 18px;
+  font-weight: 300;
+`;
+
+export const HeadingBold = styled.span`
+  font-weight: 600;
+`;
